Refactor WebSocket Server Code
