+++ platformio.ini
@@ -29,6 +29,7 @@ build_flags =
 	-D CORE_DEBUG_LEVEL=5
     ; Move all networking stuff to the protocol core 0 and leave business logic on application core 1
     -D CONFIG_ASYNC_TCP_RUNNING_CORE=0
+    -D ESP32SVELTEKIT_RUNNING_CORE=0
 	-D NO_GLOBAL_ARDUINOOTA
     ; Uncomment PROGMEM_WWW to enable the storage of the WWW data in PROGMEM
     -D PROGMEM_WWW
