+++ src/LightStateService.cpp
@@ -24,14 +24,14 @@ LightStateService::LightStateService(AsyncWebServer *server,
                                                                                                          LIGHT_SETTINGS_ENDPOINT_PATH,
                                                                                                          securityManager,
                                                                                                          AuthenticationPredicates::IS_AUTHENTICATED),
-                                                                                           _mqttPubSub(LightState::haRead, LightState::haUpdate, this, mqttClient),
-                                                                                           _webSocket(LightState::read,
-                                                                                                      LightState::update,
-                                                                                                      this,
-                                                                                                      server,
-                                                                                                      LIGHT_SETTINGS_SOCKET_PATH,
-                                                                                                      securityManager,
-                                                                                                      AuthenticationPredicates::IS_AUTHENTICATED),
+                                                                                           _mqttPubSub(LightState::homeAssistRead, LightState::homeAssistUpdate, this, mqttClient),
+                                                                                           _webSocketServer(LightState::read,
+                                                                                                            LightState::update,
+                                                                                                            this,
+                                                                                                            server,
+                                                                                                            LIGHT_SETTINGS_SOCKET_PATH,
+                                                                                                            securityManager,
+                                                                                                            AuthenticationPredicates::IS_AUTHENTICATED),
                                                                                            _mqttClient(mqttClient),
                                                                                            _lightMqttSettingsService(lightMqttSettingsService)
 {
